Add files via upload
Types of Interest
Interest is a fee for borrowing money. It is usually a percentage of the amount borrowed or saved. When you borrow money, the lender will charge you interest. When you save money, the bank will pay you interest. There are 2 types of interest, simple and compound.

If user selects "Investment", the user will be required to input:

the amount of money that they are depositing
- interest rate as a percentage
- number of years
- compound or simple interest 

Simple Interest
Simple interest is a type of interest where the interest is calculated only on the initial principal amount, without taking into account the accumulated interest from previous periods. This means that the interest does not earn interest and the interest amount remains the same for the life of the loan or savings account.

Compound Interest
Compound interest is where the interest is calculated on the initial principal amount, but also on the accumulated interest from previous periods. This means that the interest earns interest, leading to a faster growth of the investment or debt. The interest is usually compounded on a regular basis, such as annually or monthly.

If the user selects "Bond", the user will be required to input:

- the present value
- interest rate as a percentage
- number of months

Bond Repayment
A home bond repayment refers to the process of paying back the principal amount of a home bond loan, plus any accumulated interest, to the lender. A home bond is a type of loan that is used to purchase a house or property. The home bond repayment is usually paid monthly over a specified loan term. The exact amount of a monthly home bond repayment will depend on the size of the loan, the interest rate, and the loan term.